+++ src/main/java/com/devonfw/quarkus/ordermanagement/domain/repo/ItemFragment.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ItemFragment {
     void deleteByOrderId(Long id);
 
-    List<ItemEntity> findByOrderId(Long id);
+    List<ItemEntity> findItemsByOrderId(Long id);
 }
